+++ account-kit/infra/src/exports/index.ts
@@ -43,6 +43,7 @@ export {
   inkSepolia,
   mekong,
   monadTestnet,
+  monadMainnet,
   openlootSepolia,
   gensynTestnet,
   riseTestnet,
@@ -54,6 +55,9 @@ export {
   bobaSepolia,
   bobaMainnet,
   edgeMainnet,
+  stableMainnet,
+  mythosMainnet,
+  worldl3devnet,
 } from "../chains.js";
 export type * from "../client/decorators/smartAccount.js";
 export { alchemyActions } from "../client/decorators/smartAccount.js";